+++ R/cnetplot.R
@@ -257,29 +257,19 @@ cnetplot.compareClusterResult <- function(x,
 ##' @return a data.frame
 ##' @noRd
 deal_gene_pie <- function(y) {
+    ## gene_pie <- as.matrix(y[,c(1,2,9)])
+    gene_pie <- tibble::as_tibble(y[,c("Cluster", "ID", "geneID")])
+    gene_pie$geneID <- strsplit(gene_pie$geneID, '/')
+    gene_pie2 <- as.data.frame(tidyr::unnest(gene_pie, cols=geneID))
 
-    #data_pie <- as.matrix(y[,c(1,2,10)])
-    gene_pie <- as.matrix(y[,c(1,2,9)])
-    gene_pie2 <- NULL
-    for(i in seq_len(dim(gene_pie)[1])){
-        gene <- unlist(strsplit(gene_pie[i, 3], "/"))
-        for(j in seq_len(length(gene))) {
-            
-            gene_pie2 <- rbind(gene_pie2,c(gene[j],gene_pie[i,1:2]))
-        }
-    }
     gene_pie2 <- unique(gene_pie2)
     genes_unique <- unique(gene_pie2[,1])
     Cluster_unique <- unique(gene_pie2[,2])
 
-    #ID_Cluster_mat <- matrix(0,length(ID_unique),length(Cluster_unique))
-    gene_Cluster_mat <- matrix(0,length(genes_unique),length(Cluster_unique))   
-    #rownames(ID_Cluster_mat) <- ID_unique
-    #colnames(ID_Cluster_mat) <- Cluster_unique
+    gene_Cluster_mat <- matrix(0, nrow = length(genes_unique), ncol = length(Cluster_unique))   
     rownames(gene_Cluster_mat) <- genes_unique
     colnames(gene_Cluster_mat) <- Cluster_unique
     
-    #ID_Cluster_mat <- as.data.frame(ID_Cluster_mat, stringAsFactors = FALSE)
     gene_Cluster_mat <- as.data.frame(gene_Cluster_mat, stringAsFactors = FALSE)
     for(i in seq_len(nrow(gene_pie2))) {
         gene_Cluster_mat[gene_pie2[i,1],gene_pie2[i,2]] <- 1